backend/services/dspy_feedback.py
"""
DSPy-based Feedback System for Human-in-the-Loop Learning.
Implements intelligent feedback processing and solution improvement.
"""

import dspy
import asyncio
import json
from typing import Dict, Any, List, Optional
from datetime import datetime

class DSPyFeedbackSystem:
    def __init__(self):
        # Initialize DSPy with a language model
        # In production, configure with your preferred LM
        try:
            # Configure DSPy (you'll need to set up your LM)
            lm = dspy.OpenAI(
                model="gpt-3.5-turbo",
                api_key="your-api-key-here",  # Replace with actual API key
                max_tokens=1000
            )
            dspy.settings.configure(lm=lm)
        except:
            # Fallback for demo without API key
            lm = None
            
        self.feedback_history = []
        self.solution_improvements = {}
        
    async def process_feedback(
        self,
        problem_id: str,
        accuracy_rating: int,
        clarity_rating: str,
        comments: Optional[str] = None
    ):
        """
        Process feedback using DSPy for intelligent learning.
        """
        feedback_data = {
            "problem_id": problem_id,
            "accuracy_rating": accuracy_rating,
            "clarity_rating": clarity_rating,
            "comments": comments,
            "timestamp": datetime.now(),
            "processed": False
        }
        
        # Store feedback
        self.feedback_history.append(feedback_data)
        
        # Process feedback if we have DSPy configured
        if hasattr(dspy.settings, 'lm') and dspy.settings.lm:
            await self._process_with_dspy(feedback_data)
        else:
            # Fallback processing without DSPy
            await self._process_feedback_fallback(feedback_data)
    
    async def _process_with_dspy(self, feedback_data: Dict[str, Any]):
        """Process feedback using DSPy optimization."""
        try:
            # Define DSPy signature for feedback analysis
            class FeedbackAnalyzer(dspy.Signature):
                """Analyze mathematical solution feedback to improve future responses."""
                feedback_rating = dspy.InputField(desc="Accuracy rating from 1-5")
                clarity_rating = dspy.InputField(desc="Clarity rating: Very Clear, Somewhat Clear, Unclear")
                comments = dspy.InputField(desc="User comments about the solution")
                improvement_suggestions = dspy.OutputField(desc="Specific suggestions to improve the solution")
                confidence_adjustment = dspy.OutputField(desc="How to adjust confidence score based on feedback")
            
            # Create and run the analyzer
            analyzer = dspy.ChainOfThought(FeedbackAnalyzer)
            
            result = analyzer(
                feedback_rating=str(feedback_data["accuracy_rating"]),
                clarity_rating=feedback_data["clarity_rating"],
                comments=feedback_data["comments"] or "No comments provided"
            )
            
            # Store the analysis results
            improvement_data = {
                "problem_id": feedback_data["problem_id"],
                "suggestions": result.improvement_suggestions,
                "confidence_adjustment": result.confidence_adjustment,
                "processed_at": datetime.now()
            }
            
            self.solution_improvements[feedback_data["problem_id"]] = improvement_data
            feedback_data["processed"] = True
            
        except Exception as e:
            print(f"DSPy processing failed: {e}")
            await self._process_feedback_fallback(feedback_data)
    
    async def _process_feedback_fallback(self, feedback_data: Dict[str, Any]):
        """Fallback feedback processing without DSPy."""
        
        # Simple rule-based improvement suggestions
        suggestions = []
        
        if feedback_data["accuracy_rating"] <= 2:
            suggestions.append("Review mathematical accuracy of the solution")
            suggestions.append("Verify calculation steps")
        
        if feedback_data["clarity_rating"] == "Unclear":
            suggestions.append("Provide more detailed explanations")
            suggestions.append("Break down complex steps into smaller parts")
            suggestions.append("Add more context for mathematical concepts")
        
        if feedback_data["comments"]:
            comments_lower = feedback_data["comments"].lower()
            if "confusing" in comments_lower:
                suggestions.append("Simplify language and explanations")
            if "wrong" in comments_lower:
                suggestions.append("Double-check mathematical calculations")
            if "incomplete" in comments_lower:
                suggestions.append("Provide more comprehensive solution steps")
        
        # Store improvement suggestions
        improvement_data = {
            "problem_id": feedback_data["problem_id"],
            "suggestions": suggestions,
            "confidence_adjustment": self._calculate_confidence_adjustment(feedback_data),
            "processed_at": datetime.now()
        }
        
        self.solution_improvements[feedback_data["problem_id"]] = improvement_data
        feedback_data["processed"] = True
    
    def _calculate_confidence_adjustment(self, feedback_data: Dict[str, Any]) -> str:
        """Calculate how to adjust confidence based on feedback."""
        
        rating = feedback_data["accuracy_rating"]
        clarity = feedback_data["clarity_rating"]
        
        if rating >= 4 and clarity == "Very Clear":
            return "Increase confidence by 10%"
        elif rating >= 3 and clarity in ["Very Clear", "Somewhat Clear"]:
            return "Maintain current confidence level"
        elif rating <= 2 or clarity == "Unclear":
            return "Decrease confidence by 20%"
        else:
            return "Decrease confidence by 10%"
    
    async def get_improvement_suggestions(self, problem_id: str) -> Optional[Dict[str, Any]]:
        """Get improvement suggestions for a specific problem."""
        return self.solution_improvements.get(problem_id)
    
    async def get_feedback_insights(self) -> Dict[str, Any]:
        """Get insights from all processed feedback."""
        
        if not self.feedback_history:
            return {"total_feedback": 0, "insights": []}
        
        total_feedback = len(self.feedback_history)
        processed_feedback = len([f for f in self.feedback_history if f["processed"]])
        
        # Calculate average ratings
        accuracy_ratings = [f["accuracy_rating"] for f in self.feedback_history]
        avg_accuracy = sum(accuracy_ratings) / len(accuracy_ratings)
        
        # Common clarity issues
        clarity_counts = {}
        for f in self.feedback_history:
            clarity = f["clarity_rating"]
            clarity_counts[clarity] = clarity_counts.get(clarity, 0) + 1
        
        # Generate insights
        insights = []
        
        if avg_accuracy < 3:
            insights.append("Overall solution accuracy needs improvement")
        
        if clarity_counts.get("Unclear", 0) > clarity_counts.get("Very Clear", 0):
            insights.append("Focus on improving explanation clarity")
        
        # Common themes in comments
        all_comments = " ".join([f["comments"] or "" for f in self.feedback_history]).lower()
        if "confusing" in all_comments:
            insights.append("Users find explanations confusing - simplify language")
        if "incomplete" in all_comments:
            insights.append("Users want more comprehensive solutions")
        
        return {
            "total_feedback": total_feedback,
            "processed_feedback": processed_feedback,
            "average_accuracy": avg_accuracy,
            "clarity_distribution": clarity_counts,
            "insights": insights,
            "improvement_suggestions_count": len(self.solution_improvements)
        }
    
    async def apply_learning_to_future_solutions(self, problem_category: str) -> List[str]:
        """Get learning-based recommendations for future solutions in a category."""
        
        # Analyze feedback for similar problems
        category_feedback = []
        for feedback in self.feedback_history:
            # In a real implementation, you'd match by problem category
            category_feedback.append(feedback)
        
        if not category_feedback:
            return ["No feedback available for this category yet"]
        
        # Generate recommendations based on past feedback
        recommendations = []
        
        low_accuracy_count = len([f for f in category_feedback if f["accuracy_rating"] <= 2])
        if low_accuracy_count > len(category_feedback) * 0.3:
            recommendations.append("Focus on mathematical accuracy for this category")
        
        unclear_count = len([f for f in category_feedback if f["clarity_rating"] == "Unclear"])
        if unclear_count > len(category_feedback) * 0.3:
            recommendations.append("Provide clearer explanations for this category")
        
        return recommendations or ["Continue with current approach - feedback is positive"]