pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai>=3.0.3",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "numpy>=2.3.2",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.1",
    "uvicorn>=0.35.0",
]
